pembuatan tablepermintaanpelanggan, tabel transaksi, merubahtabel pemesanan user- relasi v0.1

+++ database/migrations/2019_07_10_071128_create_pemesanan_users_table.php
@@ -18,7 +18,7 @@ class CreatePemesananUsersTable extends Migration
             $table->string('nama');
             $table->string('instansi');
             $table->string('alamat');
-            $table->integer('ntelp');
+            $table->double('ntelp');
             $table->string('contohygdianalisis');
             $table->string('unsurygdianalisis');
             $table->string('jml_contoh');
